Revert to real project name when exiting renaming
When you press Esc to exit while editing a project's name, the UI
won't display the unfinished project name and will just revert back
to original project name. Not doing so will otherwise cause the app
to attempt to rename a project with the unfinished name, which probably
won't exist and throws an error.