Fix activeProject flag being assigned as a project
Fix activeProject flag being accounted as a project on page load when
its value in localStorage is empty. It's a conditional logic bug which
checks for activeProject flag having some value and if it doesn't, it'll
store it as a project. Fix it to just check for the existence of
activeProject flag in localStorage and assign the variable
activeProjectName as undefined if the flag doesn't have any value.