Make ToDoSaver objects reflect the localStorage
A ToDoSaver object exists as a field for each ToDo. But when one adds
a ToDo object by calling Project.addToDo(). ToDoSaver field was set
with the state of the localStorage before adding a ToDo. Just make it
so that ToDoSaver attribute is set only after localStorage is up-to-date
with the user input data.