Prevent Enter from canceling submitted text on UI
Pressing Enter key will submit the text you've edited but removing
contenteditable attribute from the edit textbox will also trigger
blur event as a side effect and thus calling its registered event
handler and rewriting the text on UI with unedited text although the
text was already edited and submitted as far as localStorage data
is concerned. So we need to remove blur event handler when the edited
text is about to be submitted. Also remove the keydown handler as a
clean up so future edits won't stack event handlers on the edit textbox.